Add Mt. Lebanon Public Meetings Spider and Test Files
I anticipate this spider will require regular maintenance in the years to come. As of this writing, the formatting of the page is highly arbitrary, with few easily-identifiable id's or classes, and it's unclear how future updates might change the layout of the page. Keep an eye on this spider.