Add Penn Hills Public Meetings Spider
This spider is relatively unique among the spiders in Cityscrapers, as it parses first an initial page to obtain links, and then generates spiders for each of those links, parsing those and returning them. As a consequence, both the spider and the spider test file are a bit different than the norm, in manners that are hopefully detailed clearly in the actual documentation.