Add pitt_partner4work spider
At the time of development, the website had only one meeting date to scrape. Consequently, it may need to be modified in the new year as multiple dates may be added.

Meeting times were not specified. A google search showed a previous meeting to be at 8:30 AM, and I emailed info@partner4work.org to confirm is this remained the case. If it is not, this spider will need to be revised.